new NOT WORKING ---- TV Fitness Center (Lexibook) [TeamEurope] Lexibook Retro TV Game Console - Frozen - 300 Games [TeamEurope]
the TV Fitness you can select games etc. but inputs aren't fully mapped, Frozen is a more enhanced NES VT clone than we support (nw)
